third_party/WebKit/LayoutTests/http/tests/media/autoplay/document-user-activation-navigation-iframe-delegation.html

<!DOCTYPE html>
<title>Test user gesture is delegated to iframes after navigation.</title>
<script>
  eventSender.mouseDown();

  if (window.testRunner) {
    testRunner.clearBackForwardList();
    testRunner.waitUntilDone();
  }

  document.location.href = 'document-user-activation-iframe-delegation-no-gesture.html';
</script>
